# apps/api/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PULL Super App API
  version: 1.0.0
  description: |
    # PULL Super App API

    Complete API for the PULL Super App platform, providing endpoints for:

    - **Authentication**: User registration, login, token refresh, and password management
    - **Trading**: Order management, portfolio tracking, and buying power
    - **Predictions**: Prediction market events, positions, and market data
    - **RWA (Real-World Assets)**: Asset browsing, ownership, and purchases
    - **Rewards**: Points system, leaderboards, and reward redemption

    ## Authentication

    Most endpoints require authentication via Bearer token. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Tokens are obtained through the `/api/auth/login` or `/api/auth/register` endpoints.

    ## Rate Limiting

    API requests are rate-limited. Current limits are returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window

    ## Error Handling

    All errors follow a consistent format with a `success: false` flag and an `error` object containing:
    - `code`: Machine-readable error code
    - `message`: Human-readable error description

    ## Request Tracking

    All responses include a `requestId` for debugging and support purposes.
  contact:
    name: PULL Support
    email: support@pull.app
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from login or register endpoints
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          enum: [false]
          description: Always false for error responses
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Invalid email address
        requestId:
          type: string
          format: uuid
          description: Request tracking ID for debugging
          example: 550e8400-e29b-41d4-a716-446655440000
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: "2024-01-15T10:30:00.000Z"
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
        - hasNextPage
        - hasPreviousPage
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number (1-indexed)
          example: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
          example: 50
        totalItems:
          type: integer
          minimum: 0
          description: Total number of items
          example: 150
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages
          example: 3
        hasNextPage:
          type: boolean
          description: Whether there is a next page
          example: true
        hasPreviousPage:
          type: boolean
          description: Whether there is a previous page
          example: false

    # ========================================
    # Auth Schemas
    # ========================================
...
    CreateOrderRequest:
      type: object
      required:
        - assetClass
        - symbol
        - side
        - type
        - quantity
      properties:
        assetClass:
          type: string
          enum: [crypto, prediction, rwa]
          description: Asset class for the order
          example: crypto
        symbol:
          type: string
          description: Trading symbol/ticker
          example: BTC-USD
        side:
          type: string
          enum: [buy, sell]
          description: Order side
          example: buy
        type:
          type: string
          enum: [market, limit, stop, stop_limit]
          description: Order type
          example: limit
        quantity:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Order quantity
          example: 0.5
        price:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Limit price (required for limit and stop_limit orders)
          example: 45000.00
        stopPrice:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Stop trigger price (required for stop and stop_limit orders)
          example: 44000.00
        timeInForce:
          type: string
          enum: [day, gtc, ioc, fok]
          default: gtc
          description: "Time in force: day, good-til-cancelled, immediate-or-cancel, fill-or-kill"
          example: gtc
        clientOrderId:
          type: string
          description: Client-provided order ID for tracking
          example: my-order-123
        expiresAt:
          type: integer
          description: Unix timestamp when the order should expire
          example: 1705331400000
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for the order
    Order:
      type: object
      required:
        - id
        - userId
        - assetClass
        - symbol
        - side
        - type
        - quantity
        - status
        - createdAt
      properties:
        id:
          type: string
          description: Unique order ID
          example: ord_abc123def456
        userId:
          type: string
          description: User ID who placed the order
          example: usr_abc123def456
        assetClass:
          type: string
          enum: [crypto, prediction, rwa]
          description: Asset class
          example: crypto
        symbol:
          type: string
          description: Trading symbol
          example: BTC-USD
        side:
          type: string
          enum: [buy, sell]
          description: Order side
          example: buy
        type:
          type: string
          enum: [market, limit, stop, stop_limit]
          description: Order type
          example: limit
        quantity:
          type: number
          description: Order quantity
          example: 0.5
        price:
          type: number
          nullable: true
          description: Limit price
          example: 45000.00
        stopPrice:
          type: number
          nullable: true
          description: Stop price
        timeInForce:
          type: string
          enum: [day, gtc, ioc, fok]
          description: Time in force
          example: gtc
        status:
          type: string
          enum: [pending, submitted, accepted, partial_fill, filled, cancelled, rejected, expired]
          description: Current order status
          example: filled
        filledQuantity:
          type: number
          description: Quantity that has been filled
          example: 0.5
        averagePrice:
          type: number
          nullable: true
          description: Average fill price
          example: 44950.00
        clientOrderId:
          type: string
          nullable: true
          description: Client-provided order ID
          example: my-order-123
        createdAt:
          type: string
          format: date-time
          description: Order creation timestamp
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:30:05.000Z"
    OrderFill:
      type: object
      properties:
        id:
          type: string
        price:
          type: number
        quantity:
          type: number
        fee:
          type: number
        timestamp:
          type: string
          format: date-time
    Portfolio:
      type: object
      required:
        - positions
        - summary
      properties:
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
          description: List of portfolio positions
        summary:
          type: object
          required:
            - totalValue
            - totalCost
            - totalUnrealizedPnL
            - totalRealizedPnL
            - totalPnLPercent
            - positionCount
          properties:
            totalValue:
              type: number
              description: Total portfolio value
              example: 125000.00
            totalCost:
              type: number
              description: Total cost basis
              example: 100000.00
            totalUnrealizedPnL:
              type: number
              description: Total unrealized profit/loss
              example: 25000.00
            totalRealizedPnL:
              type: number
              description: Total realized profit/loss
              example: 5000.00
            totalPnLPercent:
              type: number
              description: Total P&L percentage
              example: 25.0
            positionCount:
              type: integer
              description: Number of open positions
              example: 5
    Position:
      type: object
      required:
        - symbol
        - assetClass
        - quantity
        - averageCost
        - currentPrice
        - marketValue
      properties:
        symbol:
          type: string
          description: Position symbol
          example: BTC-USD
        assetClass:
          type: string
          enum: [crypto, prediction, rwa]
          description: Asset class
          example: crypto
        quantity:
          type: number
          description: Position quantity
          example: 1.5
        averageCost:
          type: number
          description: Average cost per unit
          example: 42000.00
        currentPrice:
          type: number
          description: Current market price
          example: 45000.00
        marketValue:
          type: number
          description: Current market value
          example: 67500.00
        unrealizedPnL:
          type: number
          description: Unrealized profit/loss
          example: 4500.00
        unrealizedPnLPercent:
          type: number
          description: Unrealized P&L percentage
          example: 7.14
    BuyingPower:
      type: object
      required:
        - available
        - held
        - pending
        - total
      properties:
        available:
          type: number
          description: Available buying power
          example: 50000.00
        held:
          type: number
          description: Amount held in open orders
          example: 10000.00
        pending:
          type: number
          description: Pending deposits/withdrawals
          example: 5000.00
        total:
          type: number
          description: Total buying power
          example: 65000.00

    # ========================================
    # Prediction Schemas
    # ========================================
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Missing or invalid authorization header
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
    ForbiddenError:
      description: Access denied to the requested resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have access to this resource
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
    NotFoundError:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: The requested resource was not found
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
  parameters:
...
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
paths:
...
  /api/v1/trading/orders:
    get:
      tags:
        - Trading
      summary: List orders
      description: Get a paginated list of user's orders with optional status filter
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by order status
          schema:
            type: string
            enum: [pending, submitted, accepted, partial_fill, filled, cancelled, rejected, expired]
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  timestamp:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Trading
      summary: Create order
      description: Place a new trading order
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/Order"
                  timestamp:
                    type: string
                    format: date-time
        "400":
          description: Invalid order or insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/trading/orders/{orderId}:
    get:
      tags:
        - Trading
      summary: Get order details
      description: Get detailed information about a specific order including fills
      operationId: getOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Order"
                      - type: object
                        properties:
                          fills:
                            type: array
                            items:
                              $ref: "#/components/schemas/OrderFill"
                  timestamp:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Trading
      summary: Cancel order
      description: Cancel an open order
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID to cancel
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/Order"
                  timestamp:
                    type: string
                    format: date-time
        "400":
          description: Order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/trading/portfolio:
    get:
      tags:
        - Trading
      summary: Get portfolio
      description: Get user's portfolio positions and summary
      operationId: getPortfolio
      security:
        - bearerAuth: []
      parameters:
        - name: assetClass
          in: query
          description: Filter by asset class
          schema:
            type: string
            enum: [crypto, prediction, rwa]
      responses:
        "200":
          description: Portfolio data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/Portfolio"
                  timestamp:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/trading/buying-power:
    get:
      tags:
        - Trading
      summary: Get buying power
      description: Get user's available buying power and held funds
      operationId: getBuyingPower
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Buying power information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/BuyingPower"
                  timestamp:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ========================================
  # Predictions Endpoints
  # ========================================
...